Details (ex Toggle) changes
- Renamed the component into Details.
- Re-enable the sliding transition.
- Improved pattern to prevent the animation on mount.
- Coding style fixes.
