+++ devbox/apps/index.js
@@ -16,6 +16,7 @@ export { default as Colors } from './Colors'
 export { default as ContextMenu } from './ContextMenu'
 export { default as DataView } from './DataView'
 export { default as DateRangePicker } from './DateRangePicker'
+export { default as Details } from './Details'
 export { default as DiscButton } from './DiscButton'
 export { default as Distribution } from './Distribution'
 export { default as DropDown } from './DropDown'
@@ -50,7 +51,6 @@ export { default as TextCopy } from './TextCopy'
 export { default as TextStyles } from './TextStyles'
 export { default as Theme } from './Theme'
 export { default as Timer } from './Timer'
-export { default as Toggle } from './Toggle'
 export { default as TokenBadge } from './TokenBadge'
 export { default as TransactionBadge } from './TransactionBadge'
 export { default as TransactionProgress } from './TransactionProgress'
