+++ src/components/index.js
@@ -25,6 +25,7 @@ export { default as CircleGraph } from './CircleGraph/CircleGraph'
 export { default as ContextMenu } from './ContextMenu/ContextMenu'
 export { default as ContextMenuItem } from './ContextMenu/ContextMenuItem'
 export { default as DateRangePicker } from './DateRangePicker/DateRangePicker'
+export { default as Details } from './Details/Details'
 export { default as Distribution } from './Distribution/Distribution'
 export { default as DropDown } from './DropDown/DropDown'
 export { default as EmptyStateCard } from './Card/EmptyStateCard'
@@ -84,4 +85,3 @@ export { default as Text } from './Text/Text'
 // Experimental components
 export { default as _AutoComplete } from './AutoComplete/AutoComplete'
 export { default as _AutoCompleteSelected } from './AutoComplete/AutoCompleteSelected'
-export { default as _Toggle } from './Toggle/Toggle'
